requirements.txt
yandexcloud
pandas>=1.5.0
python-dotenv>=0.19.0
requests>=2.28.0
grpcio>=1.48.0
protobuf>=3.19.0
tqdm>=4.64.0